jira-consumer/requirements.txt
kafka-python==2.0.2
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.0
tenacity==8.2.3
python-dotenv==1.0.0
prometheus-client==0.19.0